Split webpack config to separate files
